// gravitee-apim-webui-libs/gravitee-dashboard/src/lib/models/dashboard.model.ts
/*
 * Copyright (C) 2025 The Gravitee team (http://gravitee.io)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *         http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { Widget } from '../components/widget/model/widget/widget.model';

export interface Dashboard {
  id: string;
  name: string;
  createdBy: string;
  createdAt: string;
  lastModified: string;
  labels: Record<string, string>;
  widgets: Widget[];
}